Finalize refactoring + fix expand/collapse all traces for terminal
