+++ src/app/api/interfaces/credentials.interface.ts
