+++ src/app/api/interfaces/trace.interface.ts
