+++ src/app/app.module.ts
@@ -1,12 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import {
-  PathLocationStrategy,
-  LocationStrategy,
-  DatePipe,
-} from '@angular/common';
+import { PathLocationStrategy, LocationStrategy } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { ServiceWorkerModule } from '@angular/service-worker';
 
 import { Angulartics2Module } from 'angulartics2';
@@ -16,47 +11,22 @@ import { LoggerModule, NgxLoggerLevel } from 'ngx-logger';
 import { AppRoutingModule } from './app-routing.module';
 import { UiModule } from './ui.module';
 import { EnvModule, environment } from './env.module';
+import { SharedModule } from './shared/shared.module';
+import { ApiModule } from './api/api.module';
 
 import { AppComponent } from './app.component';
-import { LoginViewComponent } from './login-view.component';
-import { SandboxesViewComponent } from './sandboxes-view.component';
-import { StackTraceComponent } from './stack-trace.component';
-import { RedirectViewComponent } from './redirect-view.component';
-import { TraceViewComponent } from './trace-view.component';
-import { TerminalViewComponent } from './terminal-view.component';
-import { DebugFormComponent } from './debug-form.component';
-import { DebugStatusApi } from './debug-status-api.service';
-import { PreferencesStorage } from './preferences-storage.service';
-import { LazyJsonComponent } from './lazy-json.component';
-import { StackFilterComponent } from './stack-filter.component';
-import { JsonViewerComponent } from './json-viewer.component';
-
-import { ScrollGlueDirective } from './scrollGlue.directive';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    LoginViewComponent,
-    SandboxesViewComponent,
-    StackTraceComponent,
-    TraceViewComponent,
-    TerminalViewComponent,
-    RedirectViewComponent,
-    DebugFormComponent,
-    LazyJsonComponent,
-    ScrollGlueDirective,
-    StackFilterComponent,
-    JsonViewerComponent,
-  ],
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    FormsModule,
     // Environement module
     EnvModule,
     // Provide Core Services
     AppRoutingModule,
     UiModule,
+    ApiModule,
     // Analytics
     Angulartics2Module.forRoot([Angulartics2GoogleTagManager]),
     // Service worker
@@ -68,13 +38,10 @@ import { ScrollGlueDirective } from './scrollGlue.directive';
       level: NgxLoggerLevel.LOG,
       serverLogLevel: NgxLoggerLevel.OFF,
     }),
+    // Shared
+    SharedModule,
   ],
-  providers: [
-    DebugStatusApi,
-    PreferencesStorage,
-    { provide: LocationStrategy, useClass: PathLocationStrategy },
-    DatePipe,
-  ],
+  providers: [{ provide: LocationStrategy, useClass: PathLocationStrategy }],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
