+++ src/app/auth/auth-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule, Routes } from '@angular/router';
+
+import { LoginComponent } from './login/login.component';
+
+export const ROUTES: Routes = [
+  {
+    path: 'login',
+    component: LoginComponent,
+  },
+];
+
+@NgModule({
+  imports: [CommonModule, RouterModule.forChild(ROUTES)],
+})
+export class AuthRoutingModule {}
