+++ src/app/auth/login/login.component.scss
@@ -1,0 +1,14 @@
+mat-card {
+  max-width: 50vw;
+  margin: auto;
+}
+mat-progress-spinner {
+  margin: auto;
+}
+mat-card-content {
+  display: flex;
+  flex-direction: column;
+}
+mat-card-content > * {
+  width: 100%;
+}
