+++ src/app/auth/login/login.component.ts
@@ -4,69 +4,19 @@ import { Router } from '@angular/router';
 
 import { NGXLogger } from 'ngx-logger';
 
-import { ENVIRONMENT_ID } from './env.module';
-import { Platform } from './environment.interface';
-import { Credentials } from './credentials.interface';
-import { getSecureUrl } from './utils';
+import { ENVIRONMENT_ID } from '../../env.module';
+import { Platform } from '../../api/interfaces/environment.interface';
+import { Credentials } from '../../api/interfaces/credentials.interface';
+import { getSecureUrl } from '../../utils';
 
 const CUSTOM_API_URL = '<custom>';
 
 @Component({
-  selector: 'zp-login-view',
-  template: `
-  <form (ngSubmit)="onSubmit(form)" novalidate #form="ngForm">
-    <mat-card *ngIf="!connecting">
-      <mat-card-title>Login</mat-card-title>
-      <mat-card-content>
-        <mat-form-field>
-          <input matInput placeholder="Username" name="username" type="text" required ngModel>
-        </mat-form-field>
-        <mat-form-field>
-          <input matInput placeholder="Password" name="password" type="password" required ngModel>
-        </mat-form-field>
-        <mat-form-field>
-          <mat-select placeholder="Api Url" name="apiUrl" required ngModel>
-            <mat-option *ngFor="let platform of platforms" [value]="platform.url">
-              {{ platform.name }}
-            </mat-option>
-          </mat-select>
-        </mat-form-field>
-        <mat-form-field *ngIf="isCustomApiUrl()">
-          <input matInput placeholder="Custom Api Url" name="customApiUrl" type="text" required ngModel>
-        </mat-form-field>
-      </mat-card-content>
-      <mat-card-actions align="end">
-        <button mat-raised-button>Connect</button>
-      </mat-card-actions>
-    </mat-card>
-    <mat-card *ngIf="connecting">
-      <mat-card-title>Connecting ...</mat-card-title>
-      <mat-card-content>
-        <mat-progress-spinner mode="indeterminate"></mat-progress-spinner>
-      </mat-card-content>
-    </mat-card>
-  </form>
-  `,
-  styles: [
-    `
-    mat-card {
-      max-width: 50vw;
-      margin: auto;
-    }
-    mat-progress-spinner {
-      margin: auto;
-    }
-    mat-card-content {
-      display: flex;
-      flex-direction: column;
-    }
-    mat-card-content > * {
-      width: 100%;
-    }
-  `,
-  ],
+  selector: 'zp-login',
+  templateUrl: 'login.component.html',
+  styleUrls: ['login.component.scss'],
 })
-export class LoginViewComponent implements OnInit {
+export class LoginComponent implements OnInit {
   platforms: Platform[];
   connecting = false;
 
@@ -81,7 +31,7 @@ export class LoginViewComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.logger.log('LoginViewComponent::ngOnInit', this.form);
+    this.logger.log('LoginComponent::ngOnInit', this.form);
   }
 
   async onSubmit({ value, valid }: { value: Credentials; valid: boolean }) {
