+++ src/app/sandboxes/sandboxes-view/sandboxes-view.component.scss
@@ -1,0 +1,10 @@
+[mat-list-icon] {
+  color: rgba(0, 0, 0, 0.54);
+}
+.mat-list {
+  width: 50%;
+  float: left;
+}
+mat-list-item {
+  cursor: pointer;
+}
