+++ src/app/shared/redirect-view/redirect-view.component.ts
