+++ src/app/traces/lazy-json/lazy-json.component.ts
