+++ src/app/traces/stack-trace/stack-trace.component.scss
@@ -1,0 +1,12 @@
+.Trace--USR {
+  color: #1b6fcb;
+  font-weight: bold;
+}
+.Trace--CMT {
+  color: #c1b8b6;
+}
+.Trace--MS,
+.Trace--ME {
+  color: #919191;
+  font-weight: bold;
+}
