+++ src/app/traces/traces-view/traces-view.component.html
@@ -1,0 +1,65 @@
+<mat-sidenav-container class="Container--Sidenav">
+  <h1>
+    <mat-icon mat-list-icon [routerLink]="['/sandboxes']">arrow_back</mat-icon>
+    <span>Trace for {{sandboxId}}</span>
+  </h1>
+  <zp-debug-form [sandboxId]="sandboxId" [services]="services"></zp-debug-form>
+  <mat-toolbar>
+    <mat-form-field class="Filter">
+    <input matInput (keyup)="onFiltering($event.target.value)" placeholder="Filter">
+    </mat-form-field>
+    <span class="Spacer"></span>
+    <button mat-icon-button>
+      <mat-icon aria-label="Clear trace list" (click)="onClearClick()">clear</mat-icon>
+    </button>
+  </mat-toolbar>
+  <mat-table #table [dataSource]="source" class="Table">
+    <!-- Ctx Column -->
+    <ng-container matColumnDef="ctx">
+      <mat-header-cell class="HeaderCell HeaderCell--Ctx" *matHeaderCellDef> Id </mat-header-cell>
+      <mat-cell class="Cell Cell--Ctx" *matCellDef="let row" (click)="onShowClick(row);sidenav.toggle()">
+        {{row.ctx}}
+      </mat-cell>
+    </ng-container>
+    <!-- Actions Column -->
+    <ng-container matColumnDef="actions">
+      <mat-header-cell class="HeaderCell HeaderCell--Actions" *matHeaderCellDef> Actions </mat-header-cell>
+      <mat-cell class="Cell Cell--Actions" *matCellDef="let row">
+        <button mat-icon-button>
+          <mat-icon aria-label="Download full trace" mat-list-icon (click)="onDownloadClick(row)">get_app</mat-icon>
+        </button>
+        <button mat-icon-button>
+          <mat-icon aria-label="Export trace as test case" mat-list-icon (click)="onExportClick(row)">bug_report</mat-icon>
+        </button>
+      </mat-cell>
+    </ng-container>
+    <!-- Ts Column -->
+    <ng-container matColumnDef="ts">
+      <mat-header-cell class="HeaderCell HeaderCell--Ts" *matHeaderCellDef> Date </mat-header-cell>
+      <mat-cell class="Cell Cell--Ts" *matCellDef="let row" (click)="onShowClick(row);sidenav.toggle()">
+        {{row.ts | date:'HH:mm:SS'}}
+      </mat-cell>
+    </ng-container>
+    <!-- Name Column -->
+    <ng-container matColumnDef="name">
+      <mat-header-cell class="HeaderCell HeaderCell--Name" *matHeaderCellDef> Name </mat-header-cell>
+      <mat-cell class="Cell Cell--Name" *matCellDef="let row" (click)="onShowClick(row);sidenav.toggle()">
+        {{row.data.name ? row.data.name : 'UNAVAILABLE'}}
+      </mat-cell>
+    </ng-container>
+    <!-- Owner Column -->
+    <ng-container matColumnDef="owner">
+      <mat-header-cell class="HeaderCell HeaderCell--Owner" *matHeaderCellDef> Owner </mat-header-cell>
+      <mat-cell class="Cell Cell--Owner" *matCellDef="let row" (click)="onShowClick(row);sidenav.toggle()">
+        {{row.owner}}
+      </mat-cell>
+    </ng-container>
+
+    <mat-header-row *matHeaderRowDef="columns"></mat-header-row>
+    <mat-row *matRowDef="let row; columns: columns;" class="Table__Row"></mat-row>
+  </mat-table>
+
+  <mat-sidenav #sidenav mode="over" position="end">
+    <zp-stack-trace [traces]="selection"></zp-stack-trace>
+  </mat-sidenav>
+</mat-sidenav-container>
